[CI] Switch away from kubernetes executor mode
We are running into reliability problems with the kubernetes executor mode,
hypothesized currently to be due to a complex interaction with
konnectivity-agent pods dying and subsequently killing in-process execs through
the k8s control plane. The plan is to switch back to the in-container executor
mode for now while we sort out the issue upstream with the konnectivity
developers.