[CI] Refactor helm charts and monitoring into module
This patch refactors the helm charts and monitoring into a terraform module.
This makes it easier to reuse them for a HA setup where we have multiple
clusters.

We leave the metrics container out on purpose as we only need to run it in
one cluster.

Pull Request: https://github.com/llvm/llvm-zorg/pull/442