[CI] Add Lifecycle Rules to GCS Buckets
This patch adds lifecycle rules to the GCS buckets to delete old objects.
We do not want to keep object files around for very long as they quickly
become out of date. GCS does not keep track of the last time a file was
fetched, so we have to rely on the creation date. The granularity for when
to delete an object can only be specified in days, so we set it to one
day for now.

Pull Request: https://github.com/llvm/llvm-zorg/pull/519